requirements.txt
python-telegram-bot>=20.7
psycopg2-binary>=2.9.9
APScheduler>=3.10.4
python-dotenv>=1.0.1
pip-audit>=2.7.0
